+++ Tests/xcodeproj-cliTests/TestResources/Sources/Models/Item.swift
@@ -1,6 +1,6 @@
 import Foundation
 
 struct Item: Identifiable {
-    let id = UUID()
-    let title: String
+  let id = UUID()
+  let title: String
 }