+++ Tests/xcodeproj-cliTests/TestResources/Sources/Views/ContentView.swift
@@ -1,7 +1,7 @@
 import SwiftUI
 
 struct ContentView: View {
-    var body: some View {
-        Text("Hello, World!")
-    }
+  var body: some View {
+    Text("Hello, World!")
+  }
 }