+++ Tests/xcodeproj-cliTests/TestResources/Sources/Views/ItemView.swift
@@ -1,9 +1,9 @@
 import SwiftUI
 
 struct ItemView: View {
-    let item: Item
-    
-    var body: some View {
-        Text(item.title)
-    }
+  let item: Item
+
+  var body: some View {
+    Text(item.title)
+  }
 }